% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{connect}
\alias{connect}
\title{Connects to the experimental collection in the research database}
\usage{
connect()
}
\arguments{
\item{q}{Query string}

\item{f}{Fields wanted in the result from the query}
}
\value{
A function to query from the connected collection
Lets you query the experimental collection from the research database
}
\description{
Connects to the experimental collection in the research database
}
