% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{connectTo}
\alias{connectTo}
\title{Connects to the specified collection in the specified database}
\usage{
connectTo(from, to)
}
\arguments{
\item{from}{(string) Database name}

\item{to}{(string) Collection name, must exist in the database to be able to query it.}

\item{q}{Query string}

\item{f}{Fields wanted in the result from the query}
}
\value{
A function to query from the connected collection
Lets you query the experimental collection from the research database
}
\description{
Connects to the specified collection in the specified database
}
