% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportDatabaseToCSV}
\alias{exportDatabaseToCSV}
\title{Exports the whole experimental collection from the research database into a CSV file.}
\usage{
exportDatabaseToCSV()
}
\description{
Exports the whole experimental collection from the research database into a CSV file.
}
\note{
the CSV file will be saved by default on the working directory of the project.
}
