% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportDataFrameToCSV}
\alias{exportDataFrameToCSV}
\title{Exports a data frame into an specified CSV file}
\usage{
exportDataFrameToCSV(df, filename)
}
\arguments{
\item{df}{data frame to be exported}

\item{filename}{Name of the file in which the data frame will be inserted}
}
\description{
Exports a data frame into an specified CSV file
}
\note{
filename can also receive a relative path to where the CSV file will be save.
Keep in mind that the path should and with <name-of-file>.csv in order for the
function to recognize it as a valid file name. In case the just <name-of-file>.csv
is provided with no explicit path to where the file should be saved, the function will
save it into the project's working directory.
}
