% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getCapPointByExp}
\alias{getCapPointByExp}
\title{Returns all capture points asociated with the specifed sessionID}
\usage{
getCapPointByExp(sessionID)
}
\arguments{
\item{sessionID}{Identifier of the session as specified in the CSV as the dataID field}
}
\value{
Returned Values: puntoCaptID, puntoCaptLatitud, puntoCaptLongitud, puntoCaptAlturaSurco, puntoCaptAmplitudSurco, puntoCaptRadioEfectivo
}
\description{
Returns all capture points asociated with the specifed sessionID
}
