% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getCapPointByLote}
\alias{getCapPointByLote}
\title{Returns all capture points within the specifed lote}
\usage{
getCapPointByLote(loteID)
}
\arguments{
\item{loteID}{Identifier of the lote as specified in the CSV}
}
\value{
Returned Values: puntoCaptID, puntoCaptLatitud, puntoCaptLongitud, puntoCaptAlturaSurco, puntoCaptAmplitudSurco, puntoCaptRadioEfectivo
}
\description{
Returns all capture points within the specifed lote
}
