% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getLamByCapHour}
\alias{getLamByCapHour}
\title{Returns all capture points within an specified hour-of-capture range from different experimental sessions}
\usage{
getLamByCapHour(sh, eh)
}
\arguments{
\item{sh}{Minimum value of the search range.}

\item{eh}{Maximum value of the search range.}
}
\value{
Returned Values: laminaID, laminaCaptureMoment, laminaPrecipitacion, fechaCaptLamina
}
\description{
Returns all capture points within an specified hour-of-capture range from different experimental sessions
}
\note{
The humidity range is inclusive. If any of the parameters ISN't
going to be used, specify it as NULL. Otherwise there'll be an
error.
}
