% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getLamByCapPoint}
\alias{getLamByCapPoint}
\title{Returns all laminas capture in the specified capture point in diferent experimental sessions}
\usage{
getLamByCapPoint(captPointID)
}
\arguments{
\item{capPointID}{Identifier of the capture point the laminas are asociated to.}
}
\value{
Returned Values: laminaID, laminaCaptureMoment, laminaPrecipitacion, fechaCaptLamina
}
\description{
Returns all laminas capture in the specified capture point in diferent experimental sessions
}
