% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getLoteByDate}
\alias{getLoteByDate}
\title{Returns all lotes within the specifed plantation date range}
\usage{
getLoteByDate(sd, ed)
}
\arguments{
\item{sd}{Starting date of the search range.}

\item{ed}{Ending date of the search range.}
}
\value{
values Returned: loteID, lotePlantationDate, loteLatitude, loteLongitude
}
\description{
Returns all lotes within the specifed plantation date range
}
\note{
The date range is inclusive. If any of the parameters ISN't
going to be used, specify it as NULL. Otherwise there'll be an
error.
}
