% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getLoteByExp}
\alias{getLoteByExp}
\title{Returns the specific lotes asociated with the specifed sessionID}
\usage{
getLoteByExp(sessionID)
}
\arguments{
\item{sessionID}{Identifier of the session as specified in the CSV as the dataID field}
}
\value{
Values Returned: loteID, lotePlantationDate, loteLatitude, loteLongitude
}
\description{
Returns the specific lotes asociated with the specifed sessionID
}
