% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getPlantByLote}
\alias{getPlantByLote}
\title{Returns all the plantones asociated with the specified loteID}
\usage{
getPlantByLote(loteID)
}
\arguments{
\item{loteID}{Identifier of the lote as specified in the CSV}
}
\value{
Returned Values: plantonID, plantonLongitude, plantonLatitude, plantonBordeBit
}
\description{
Returns all the plantones asociated with the specified loteID
}
