% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getPlantCapPoint}
\alias{getPlantCapPoint}
\title{Returns all the capture points asociated with the loteID and plantonID specified}
\usage{
getPlantCapPoint(loteID, plantonID)
}
\arguments{
\item{loteID}{Identifier of the lote as specified in the CSV}

\item{platonID}{Identifier of the planton asociated to the loteID}
}
\value{
Returned Values: puntoCaptID, puntoCaptLatitud, puntoCaptLongitud, puntoCaptAlturaSurco, puntoCaptAmplitudSurco, puntoCaptRadioEfectivo
}
\description{
Returns all the capture points asociated with the loteID and plantonID specified
}
