% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getSessionByDate}
\alias{getSessionByDate}
\title{Returns all experimental sessions within the date range provided}
\usage{
getSessionByDate(sd, ed)
}
\arguments{
\item{sd}{Starting date of the search range.}

\item{ed}{Ending date of the search range.}
}
\value{
Values Returned: dataID, windDirection, windSpeed, temperature, humidity, startDate, endDate
}
\description{
Returns all experimental sessions within the date range provided
}
\note{
The date range is inclusive. If any of the parameters ISN't
going to be used, specify it as NULL. Otherwise there'll be an
error.
}
