% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getSessionByHum}
\alias{getSessionByHum}
\title{Returns all experimental sessions wihtin the humidity range specified}
\usage{
getSessionByHum(sh, eh)
}
\arguments{
\item{sh}{Minimum value of the search range.}

\item{eh}{Maximum value of the search range.}
}
\value{
Values Returned: dataID, windDirection, windSpeed, temperature, humidity, startDate, endDate
}
\description{
Returns all experimental sessions wihtin the humidity range specified
}
\note{
The humidity range is inclusive. If any of the parameters ISN't
going to be used, specify it as NULL. Otherwise there'll be an
error.
}
