% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getSessionByID}
\alias{getSessionByID}
\title{Returns the experimental session asociated with the ID provided}
\usage{
getSessionByID(sessionID)
}
\arguments{
\item{sessionID}{Identifier of the session as specified in the CSV as the dataID field}
}
\value{
Values Returned: dataID, windDirection, windSpeed, temperature, humidity, startDate, endDate
}
\description{
Returns the experimental session asociated with the ID provided
}
