% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getSessionByLote}
\alias{getSessionByLote}
\title{Returns all experimental sessions related to the loteID provided}
\usage{
getSessionByLote(loteID)
}
\arguments{
\item{loteID}{Identifier of the lote as specified in the CSV file.}
}
\value{
values Returned: dataID, windDirection, windSpeed, temperature, humidity, startDate, endDate
}
\description{
Returns all experimental sessions related to the loteID provided
}
