% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getSessionByTemp}
\alias{getSessionByTemp}
\title{Returns all experimental sessions within the temperature range specified}
\usage{
getSessionByTemp(st, et)
}
\arguments{
\item{st}{Minimum value of the search range.}

\item{et}{Maximum value of the search range.}
}
\value{
Values Returned: dataID, windDirection, windSpeed, temperature, humidity, startDate, endDate
}
\description{
Returns all experimental sessions within the temperature range specified
}
\note{
The temperature range is inclusive. If any of the parameters ISN't
going to be used, specify it as NULL. Otherwise there'll be an
error.
}
