% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importAll}
\alias{importAll}
\title{Imports all CSV files from a given directory into the database.}
\usage{
importAll(dirpath)
}
\arguments{
\item{dirpath}{Absolute path to the directory containing the CSV files}
}
\description{
Imports all CSV files from a given directory into the database.
}
