% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importIntoCollection}
\alias{importIntoCollection}
\title{Import a dataset into an specefic collection}
\usage{
importIntoCollection(colname, filepath)
}
\arguments{
\item{colname}{Name of the collection}

\item{filepath}{Absolute path of the CSV file containing the data}
}
\description{
Import a dataset into an specefic collection
}
