% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importIntoDatabase}
\alias{importIntoDatabase}
\title{Imports from a CSV file to the mongo database.}
\usage{
importIntoDatabase(filepath)
}
\arguments{
\item{filepath}{Absolute path of for the CSV file containing the data to be inserted}
}
\description{
Imports from a CSV file to the mongo database.
}
