% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{optimizeQ}
\alias{optimizeQ}
\title{This function is intended to optimize the functional dependency of the discharge and the
time t using the squared deviation.}
\usage{
optimizeQ(Aef, Aint, hm, hd)
}
\arguments{
\item{Aef}{Efective section of the discharge.}

\item{Aint}{Efective interception area.}

\item{hm}{Maximum lamina the produces a discharge.}

\item{hd}{Value of the lamina that triggers the discharge in the container.}
}
\value{
A function with the entry parameter being a data frame as specified in the model
function.
}
\description{
This function is intended to optimize the functional dependency of the discharge and the
time t using the squared deviation.
}
\note{
The function returned automatically makes a plot comparing the observed data
with the estimate data calculated by the model function.
}
